+++ SphWaterfall/SphWaterfall/src/visualization/VisualizationManager.cpp
@@ -49,7 +49,7 @@ void VisualizationManager::renderFramesDistributed(string inputFileName, int ran
 		for (int i = 1; i < world_size; i++) {
 			unsigned int buf[1] =
 			{
-				frameParticles.size()
+				static_cast<unsigned int>(frameParticles.size())
 			};
 			MPI_Send(buf, 1, MPI_UNSIGNED, i, 0, MPI_COMM_WORLD);
 		}
@@ -61,7 +61,7 @@ void VisualizationManager::renderFramesDistributed(string inputFileName, int ran
 				//Send frame size
 				unsigned int buf[1] =
 				{
-					frame.size()
+					static_cast<unsigned int>(frame.size())
 				};
 				MPI_Send(buf, 1, MPI_UNSIGNED, target, 0, MPI_COMM_WORLD);
